refactor: separate header/footer components from index file
